Improve sub intake

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/system/hardware/IntakeSubsystem.java
@@ -294,7 +294,7 @@ public class IntakeSubsystem
     {
         return side;
     }
-    public boolean slideReached(int slideTarget)
+    public boolean slideReached(double slideTarget)
     {
         return Math.abs(inchesToTicksSlidesMotor(slideTarget) - slidePosition) < slideThreshold;
     }
